Reduce code duplicated from char.c
