+++ villain/Makefile
@@ -12,16 +12,16 @@ endif
 %.run: %.o runtime.o
 	gcc runtime.o $< -o $@ $(libs) -lm
 
-runtime.o: main.o char.o io.o symbol.o str.o math.o list.o
-	ld -r main.o char.o io.o symbol.o str.o list.o math.o -o runtime.o
+runtime.o: main.o char.o io.o symbol.o str.o math.o list.o utf8.o
+	ld -r main.o char.o io.o symbol.o str.o list.o math.o utf8.o -o runtime.o
 
 main.o: main.c types.h runtime.h
 	gcc -fPIC -c main.c -o main.o
 
-char.o: char.c types.h
+char.o: char.c types.h utf8.h
 	gcc -fPIC -c char.c -o char.o
 
-io.o: io.c runtime.h
+io.o: io.c runtime.h utf8.h
 	gcc -fPIC -c io.c -o io.o
 
 symbol.o: symbol.c str.h
@@ -29,6 +29,9 @@ symbol.o: symbol.c str.h
 
 str.o: str.c
 	gcc -fPIC -c str.c -o str.o
+
+utf8.o: utf8.c
+	gcc -fPIC -c utf8.c -o utf8.o
 
 %.o: %.s
 	nasm -f $(format) -o $@ $<
