+++ villain/io.c
@@ -2,57 +2,7 @@
 #include <inttypes.h>
 #include "types.h"
 #include "runtime.h"
-
-
-// TODO: De-duplicate between here an char.c. I can have char.c alloc a buffer
-// that's passed over here and then printf the buffer.
-
-// Output utf8 encoding of a codepoint into buffer.
-// Return the number of bytes placed in buffer.
-int utf8_encode_char(int64_t v, char *buffer) {
-  int64_t codepoint = v >> char_shift;
-  // Output to buffer using UTF-8 encoding of codepoint
-  if (codepoint < 128) {
-    buffer[0] = (char) codepoint;
-    return 1;
-  } else if (codepoint < 2048) {
-    buffer[0] = (char)(codepoint >> 6) | 192;
-    buffer[1] = ((char)codepoint & 63) | 128;
-    return 2;
-  } else if (codepoint < 65536) {
-    buffer[0] = (char)(codepoint >> 12) | 224;
-    buffer[1] = ((char)(codepoint >> 6) & 63) | 128;
-    buffer[2] = ((char)codepoint & 63) | 128;
-    return 3;
-  } else {
-    buffer[0] = (char)(codepoint >> 18) | 240;
-    buffer[1] = ((char)(codepoint >> 12) & 63) | 128;
-    buffer[2] = ((char)(codepoint >> 6) & 63) | 128;
-    buffer[3] = ((char)codepoint & 63) | 128;
-    return 4;
-  }
-}
-
-void utf8_encode_string(int64_t *str, char *buffer) {
-  int64_t len = (str[0] >> int_shift);
-  int temp;
-  int i, j;
-  int n = (len % 3 == 0) ? len / 3 : (len / 3 + 1);
-  for (i = 1; i < n; i++) {
-    for (j = 0; j < 3; j++) {
-      temp = str[i] >> (j * 21);
-      int codepoint_len = utf8_encode_char(temp, buffer);
-      buffer += codepoint_len;
-    }
-  }
-  i = (len % 3 == 0) ? 3 : (len % 3);
-  for (j = 0; j < i; j++){
-    temp = str[n] >> (j * 21);
-    int codepoint_len = utf8_encode_char(temp, buffer);
-    buffer += codepoint_len;
-  }
-  *buffer = 0;
-}
+#include "utf8.h"
 
 #define untag_port(p) ((int64_t*) (p ^ port_type_tag))
 #define port_file(p) ((FILE*) p[0])
